remove the CertifyAndSubmit component; disable API endpoints
