+++ api/routes/apds/status/put.js
@@ -7,35 +7,36 @@ module.exports = app => {
     '/apds/:id/status',
     can('submit-federal-response'),
     loadApd(),
-    async (req, res) => {
-      const apd = req.meta.apd;
+    async (_, res) => {
+      res.status(501).end();
+      // const apd = req.meta.apd;
 
-      if (apd.get('status') === 'draft') {
-        return res
-          .status(400)
-          .send({
-            error: 'apd-in-draft'
-          })
-          .end();
-      }
+      // if (apd.get('status') === 'draft') {
+      //   return res
+      //     .status(400)
+      //     .send({
+      //       error: 'apd-in-draft'
+      //     })
+      //     .end();
+      // }
 
-      const acceptableStatuses = [
-        'in review',
-        'state response',
-        'in clearance',
-        'approved',
-        'denied'
-      ];
+      // const acceptableStatuses = [
+      //   'in review',
+      //   'state response',
+      //   'in clearance',
+      //   'approved',
+      //   'denied'
+      // ];
 
-      if (!acceptableStatuses.includes(req.body.status)) {
-        return res
-          .status(400)
-          .send({ error: 'apd-invalid-status' })
-          .end();
-      }
+      // if (!acceptableStatuses.includes(req.body.status)) {
+      //   return res
+      //     .status(400)
+      //     .send({ error: 'apd-invalid-status' })
+      //     .end();
+      // }
 
-      await req.meta.apd.set({ status: req.body.status }).save();
-      return res.status(204).end();
+      // await req.meta.apd.set({ status: req.body.status }).save();
+      // return res.status(204).end();
     }
   );
 };
